pyproject.toml
[tool.poetry]
name = "gothic-zoologist"
version = "0.1.0"
description = ""
authors = ["Dominik Purgał <amonsir29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jax = {extras = ["cpu"], version = "^0.4.38"}
flax = "^0.10.2"
orbax-checkpoint = "^0.10.3"
dm-pix = "^0.4.3"
pillow = "^11.0.0"
matplotlib = "^3.10.0"
notebook = "^7.3.2"

# CIFAR 10 Dataset
# scikit-learn = "^1.6.0"
# pandas = "^2.2.3"
# torch = "^2.5.1"
# torchvision = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
